+++ doc/changelog.rst
@@ -69,6 +69,8 @@ Guide
 Importer
 ~~~~~~~~
 
+- Add Customer Importer
+- Add Product Importer
 - Add Importer
 
 Regions